Adapt fix: Build inconsistently reports errors + null root fix Problem: Building a DOM sometimes had build errors, but would not have any messages. Since messages.length is the ONLY indicator of a build error from build(), that's a problem. And the build data recorder and messages didn't tend to match in terms of what errors were reported to each.
The solution is to move to a common error reporting function called
error() in BuildResults. This creates a message, logs to build data
recorder and sets buildErr all in one place.

Additionally, there was an existing latent bug where the build was
marked as a failure internally with buildErr when the root element built to
null. But that didn't cause problems because there was no message, so the
build appeared to succeed. So that fix is in this commit too.