adapt: Add multiple history commits per deploy Previously, the history entry was committed prior to plugin actions. But now, plugins continue to need access to (and may modify) the dataDir during the act phase of deploy. So now we capture the pre-act state and a post-act (success or failure) state.
Also fixed an issue with local history dataDir path naming: colons
should not be allowed in file paths leading to dataDir. That causes
problems when the file path is used in a unix PATH-like list, which is
colon-separated. There is no provision for escaping a colon contained
within the file path.