Fix: Failed plugin actions do not create deployment on deploy:create Fixes #27 Issue: If a deployment gets to the act phase, it may be partially deployed, so the user needs access to the deployment and deployID.
createDeployment was always destroying the deployment on error, so we
now check to see if there is a history commit before destroying. A
history commit is only done right before side effects of the deployment
operation are possible. Also had to add the deployID field to the
DeployError API response.

In the CLI, the deployID is now shown in the error output.
But there was an additional problem with how we were handling
errors with dynamic tasks. If any of the dynamic tasks got a failed
message event, we were immediately throwing and then weren't waiting for
the createDeployment promise to resolve, where more complete info on
errors is present. This prevented getting backtraces on API-related
errors. All dynamic tasks had this issue, so fixed for all.