fix: Globally installed CLI errors unless in already installed project dir Fixes: #128 There were actually 3 things preventing some basic adapt CLI commands like help and list from running in a global install without a current project directory:
1. Unable to load @oclif/plugin-help
Until we're able to upstream our changes in
@unboundedsystems/plugin-help, we have to either also create our own
fork of @oclif/command or convince it to load our plugin-help instead of
@oclif/plugin-help. When running adapt out of our source repo, we were
using patch-package, but that doesn't work when actually doing a real
install.
The ugly fix is to monkeypatch using proxyquire to replace plugin-help
globally. Also backed out the patch-package changes.

2. Unable to load reanimate.ts
Reanimate does some initialization of the MummyRegistry at load time and
was expecting to always be in a project directory. The real fix should
be to completely delay reanimate init until a project directory has been
identified, but those changes are somewhat extensive. So the temporary
fix is to allow init to happen using info from the node_modules
directory where @adpt/core is.

3. Unable to load typescript
The CLI should not directly require the typescript library to be
installed in a non-dev environment. *Only* when operations require
typescript, should the version present in the project be loaded.
However, that wasn't happening. The fix is to lazy load typescript in
core.